+++ database/migrations/2025_11_30_141725_create_bom_lines_table.php
@@ -9,23 +9,21 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-   public function up()
-{
-   Schema::create('bom_lines', function (Blueprint $table) {
-    $table->id();
-    $table->unsignedBigInteger('bom_id');
-    $table->unsignedBigInteger('raw_material_id');
-    $table->float('quantity');
-    $table->decimal('cost', 12, 2);
-    $table->decimal('subtotal', 12, 2);
-    $table->timestamps();
+    public function up()
+    {
+        Schema::create('bom_lines', function (Blueprint $table) {
+            $table->id();
+            $table->unsignedBigInteger('bom_id');
+            $table->unsignedBigInteger('raw_material_id');
+            $table->decimal('quantity', 10, 2);      // UBAH dari float
+            $table->decimal('cost', 15, 2);          // sudah benar
+            $table->decimal('subtotal', 15, 2);      // sudah benar
+            $table->timestamps();
 
-    $table->foreign('bom_id')->references('id')->on('boms')->onDelete('cascade');
-    $table->foreign('raw_material_id')->references('id')->on('raw_materials')->onDelete('cascade');
-});
-
-}
-
+            $table->foreign('bom_id')->references('id')->on('boms')->onDelete('cascade');
+            $table->foreign('raw_material_id')->references('id')->on('raw_materials')->onDelete('cascade');
+        });
+    }
 
     /**
      * Reverse the migrations.
